+++ CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- 
+- Un-needed logging statement for channel defaults.
 
 ## [1.1.0] - 2025-09-09